add check for exisiting functions

+++ delete_get_results_lambda.sh
@@ -1,2 +1,6 @@
-aws lambda delete-function --function-name votingGetResults
-rm get_results_lambda.zip
+if aws lambda delete-function --function-name votingGetResults >/dev/null 2>&1 ; then
+    echo "DONE"
+    rm get_results_lambda.zip
+else
+    echo "Function does not exist"
+fi
