+++ delete_post_vote_lambda.sh
@@ -1,2 +1,6 @@
-aws lambda delete-function --function-name votingPostVote
-rm post_vote_lambda.zip
+if aws lambda delete-function --function-name votingPostVote >/dev/null 2>&1 ; then
+    echo "DONE"
+    rm post_vote_lambda.zip
+else
+    echo "Function does not exist"
+fi
